% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_list_boards.R
\name{nc_list_boards}
\alias{nc_list_boards}
\title{List details of all boards}
\usage{
nc_list_boards(archived = FALSE, deleted = FALSE)
}
\arguments{
\item{archived}{a \code{logical} of length 1. If \code{TRUE} returns also archived
boards. Default is \code{FALSE}.}

\item{deleted}{a \code{logical} of length 1. If \code{TRUE} returns also deleted
boards. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{id}: the identifier of the board
\item \code{title}: the name of the board
\item \code{owner}: the owner of the board
\item \code{color}: the color of the board
\item \code{archived}: is the board archived?
\item \code{deleted}: is the board deleted?
\item \code{n_stacks}: the number of stacks in the board
\item \code{n_labels}: the number of labels in the board
}
}
\description{
Performs an HTTP request (GET method) at the endpoint \verb{/boards} of the
Nextcloud Deck API to retrieve details of all Deck boards (including deleted
and archived boards).
}
\examples{
\dontrun{
## List details of all Nextcloud boards
nc_list_boards()

## List details of all Nextcloud boards (including archived ones)
nc_list_boards(archived = TRUE)

## List details of all Nextcloud boards (including deleted ones)
nc_list_boards(deleted = TRUE)
}
}
