% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_list_stacks.R
\name{nc_list_stacks}
\alias{nc_list_stacks}
\title{List details of all stacks for a board}
\usage{
nc_list_stacks(board_id)
}
\arguments{
\item{board_id}{a \code{integer} of length 1. The identifier of the board. Use
\code{nc_list_boards()} to get this value.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{board_id}: the identifier of the board
\item \code{id}: the identifier of the stack
\item \code{title}: the name of the stack
\item \code{n_cards}: the number of cards in the stack
}
}
\description{
Performs an HTTP request (GET method) at the endpoint
\verb{/boards/\{boardId\}/stacks} of the Nextcloud Deck API to retrieve details of
all stacks of an active Deck board.
}
\examples{
\dontrun{
## List details of all Nextcloud stacks of a board
nc_list_stacks()
}
}
