% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdeckr-package.R
\docType{package}
\name{ncdeckr-package}
\alias{ncdeckr}
\alias{ncdeckr-package}
\title{ncdeckr: An R Client to the Nextcloud Deck API}
\description{
Manages (lists, creates, deletes, and updates) boards, stacks, cards and labels of a Nextcloud Deck instance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/ncdeckr}
  \item Report bugs at \url{https://github.com/frbcesab/ncdeckr/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
