% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_list_stacks.R
\name{nc_list_stacks}
\alias{nc_list_stacks}
\title{List details of all stacks for a board}
\usage{
nc_list_stacks(board_id, deleted = FALSE)
}
\arguments{
\item{board_id}{a \code{integer} of length 1. The identifier of the board. Use
\code{nc_list_boards()} to get this value.}

\item{deleted}{a \code{logical} of length 1. If \code{TRUE} returns also deleted
stacks. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{board_id}: the identifier of the board
\item \code{id}: the identifier of the stack
\item \code{title}: the name of the stack
\item \code{deleted}: is the stack deleted?
\item \code{n_cards}: the number of cards in the stack
}
}
\description{
Performs an HTTP request (GET method) at the endpoint
\verb{/boards/\{boardId\}/stacks} of the Nextcloud Deck API to retrieve details of
all stacks of a Deck board (including deleted and archived stacks).
}
\examples{
\dontrun{
## List details of all Nextcloud boards
nc_list_stacks()

## List details of all Nextcloud boards (including deleted ones)
nc_list_stacks(deleted = TRUE)
}
}
